# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gun Del Sol API",
    "description": "High-performance async API for Solana token analysis (Modular)",
    "version": "2.0.0"
  },
  "paths": {
...
    "/wallets/{wallet_address}/tags": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Wallet Tags",
        "description": "Get tags for a wallet",
        "operationId": "get_wallet_tags_wallets__wallet_address__tags_get",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletTagsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Add Wallet Tag",
        "description": "Add a tag to a wallet",
        "operationId": "add_wallet_tag_wallets__wallet_address__tags_post",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "summary": "Remove Wallet Tag",
        "description": "Remove a tag from a wallet",
        "operationId": "remove_wallet_tag_wallets__wallet_address__tags_delete",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveTagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/wallets/batch-tags": {
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Batch Wallet Tags",
        "description": "Get tags for multiple wallets in one query",
        "operationId": "get_batch_wallet_tags_wallets_batch_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTagsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tags/{tag}/wallets": {
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "Get Wallets By Tag",
        "description": "Get all wallets with a specific tag",
        "operationId": "get_wallets_by_tag_tags__tag__wallets_get",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddTagRequest": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "is_kol": {
            "type": "boolean",
            "title": "Is Kol",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "AddTagRequest"
      },
...
      "BatchTagsRequest": {
        "properties": {
          "addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Addresses"
          }
        },
        "type": "object",
        "required": [
          "addresses"
        ],
        "title": "BatchTagsRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Simple message response"
      },
...
      "RemoveTagRequest": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "RemoveTagRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletTag": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "is_kol": {
            "type": "boolean",
            "title": "Is Kol"
          }
        },
        "type": "object",
        "required": [
          "tag",
          "is_kol"
        ],
        "title": "WalletTag"
      },
      "WalletTagsResponse": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/WalletTag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "WalletTagsResponse"
      }
    }
  }
}